backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
boto3==1.34.0
botocore==1.34.0
bcrypt==4.1.2
cryptography==41.0.7
python-dotenv==1.0.0
gunicorn==21.2.0
reportlab==4.1.0
requests>=2.31.0
# Optional: only needed if AI_MODE=gemini
google-genai>=1.0.0
